Fix setup procedure to allow integration of C routines from word2vec
Before the slower word2vec Python routines were used.
Notice: this very specific for my use case, which is: use this package
under Python2.7 on Windows.
This implies avoid using distutils in the setup file to allow the
automatic choice of a suitable C compiler by the system. A different
path could be to use distutils and Cython (possible TODO).